refactor(components): [slider] use JSX in Unit test (#8225)
